% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanSweep_Analysis.R
\name{PanSweep_Analysis}
\alias{PanSweep_Analysis}
\title{PanSweep Analysis:}
\usage{
PanSweep_Analysis(Json_Config_Path, Co_occurrence_lower_limit = 3)
}
\arguments{
\item{Json_Config_Path}{Path to the completed json config file}

\item{Co_occurrence_lower_limit}{lower bound for number of genes needed for
co-occurrence analysis. Set to 3.}
}
\value{
Returned a date stamped folder called PanSweep_Analysis_Output_YYYY-MM-DD
containing the file "PanSweep_Analysis_Output.rds".
}
\description{
This function run the analysis of MIDAS2 data and outputs a save file to be
used with the PanSweep_Shiny() function. The inputs are Json_Config_Path and
Co_occurrence_lower_limit, which is set to 3. Refer to the PanSweep github for
the Json file and explanations.
}
\details{
To run the function:
PanSweep_Analysis(Json_Config_Path = "Path/To/Json/file.json")

To change the lower bound for number of genes needed for co-occurence
analysis:
PanSweep_Analysis(Json_Config_Path = "Path/To/Json/file.json",
Co_occurrence_lower_limit = 5)
}
