% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanSweep_Shiny.R
\name{PanSweep_Shiny}
\alias{PanSweep_Shiny}
\title{PanSweep Shiny Application:}
\usage{
PanSweep_Shiny(loadData_Path)
}
\arguments{
\item{loadData_Path}{Path to the PanSweep_Analysis_Output.rds file from the
PanSweep_Analysis function.}
}
\value{
This will not return any files but run a shiny UI for visual analysis
of results.
}
\description{
This is the function to run the shiny analysis of the output from PanSweep_Analysis()
function. The file to be imputed is the "PanSweep_Analysis_Output.rds" located
in the PanSweep_Analysis_Output_YYYY-MM-DD folder created by the PanSweep_Analysis()
function.
}
\details{
To run the application use the below function:

PanSweep_Shiny(loadData_Path = "Path/to/PanSweep_Analysis_Output_YYYY-MM-DD/PanSweep_Analysis_Output.rds")

Note: The PanSweep_Analysis_Output.rds is not required to be in the folder
called PanSweep_Analysis_Output_YYYY-MM-DD for the application to be run

Once the function is called a shiny application will run and the R console cannot
be used until the application is closed.
}
